video: tegra: nvmap: Use copy_from_user to cache data
A user buffer may be paged out while memcpy() does not have a
capability to handle page faults like copy_{from,to}_user().

This page fault may cause a kernel panic due to data abort:
    Unable to handle kernel paging request at virtual address 2fad123000
    Mem abort info:
      ESR = 0x9600000b
      Exception class = DABT (current EL), IL = 32 bits
      SET = 0, FnV = 0
      EA = 0, S1PTW = 0
    Data abort info:
      ISV = 0, ISS = 0x0000000b
      CM = 0, WnR = 0
    [0000002fad123000] *pgd=000000012655f003, *pud=000000012655f003,
                       *pmd=000000010abb1003, *pte=0068000110424bd3
    Internal error: Oops: 9600000b [#1] PREEMPT SMP
    PC is at __memcpy_toio+0x6c/0xc0
    LR is at nvmap_ioctl_rw_handle+0x518/0x9c8
    x21: 0000002fad123000
    x1 : 0000002fad123000
    Kernel panic - not syncing: Fatal exception

Note that the kernel crash log doesn't really show that a page
out according to the PTE value, as the PTE value was dumped at
CPU data abort routine which might not accurately reflect the
status where this page out happened. Yet, it's well-known that
memcpy() doesn't have the protection for page-out. Therefore,
a solution here is to use copy_from_user() instead.

The reason of using memcpy_toio() here was to handle a target
address in the device space. So this patch keeps the memcpy()
call but adds an extra copy_from_user() to cache the data.

Note that adding a copy_from_user() caching may increase some
memcpy overhead, but it's a necessary compromise to choose it
over the performance due to the statbility concern.

Bug 200513687

Change-Id: I9edba9f74da50c3404c5e2bef183416fa84545e5
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2156583
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: Manish Tuteja <mtuteja@nvidia.com>