Revert "video: tegra: dc: add hdcp delay for LG model"
This reverts commit 01c261f289ea9ff65b3933d59638e6ffc86b5c21.

There have been a lot of reported issues for the LG 4K TVs with respect
to modesets blanking the display. As a WAR, we added the quirk for these
TVs where we delay the HDCP and mitigate the blanking.

Since we've found the potential root cause of the modeset blanking
issue in I37ec337fafb5663816afbcbad0a5956c81d3c919, this change removes
the LG 4K TVs from the WAR list.

Bug 200507605
Bug 200536083
Bug 2671778

Change-Id: I8e99a7f3a8ffb82b492f2bd61430b8892fefeb96
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2181561
Reviewed-by: Manish Tuteja <mtuteja@nvidia.com>
Tested-by: Manish Tuteja <mtuteja@nvidia.com>