Thermal: pwm-fan: Reset pwm on resume if inverted
Some boards share the same regulator with fan and it can set the fan to 100% with inverted PWM.
In such cases we need to enable PWM and set duty back to 0.
pwm_config checks current state and does not allow setting the same value.
Bypass the check by setting an arbitrary value and then setting it back to 0%.