t210: la: Fix latency allowance
T210 ram bus width is 64bit, unlike T210B01 that is 128bit

Correct that to fix bandwidth starvation issues with clients.

TODO:
In the future, the driver should check if SoC is t210 or t210b01 and set that properly.
On the other hand reporting lower bandwidth for latency allowance doesn't have side effects.