video: tegra: dc: improve dsi backlight experience
When a dsi display is powered down, the backlight most of the times stays on.

That also happens in suspend, where it gets disabled by the suspend routine
of the backlight driver.

All panels produce artifacts during power down sequence.
So in order to improve user experience, disable the backlight from dsi driver
and restore it during enable.

The function called is the one that is called during a shutdown
that only disables the backlight.
That avoids changing the brightness prop which is used by userspace for load/restore.
It also simplifies the call to a single update during enable/resume.